+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.0.2+4
+
+* Fix: nop()の制御を確実にイテレーションされる方法に変更
+
+## 0.0.2+3
+
+* Fix: Streams.generate()の内部構造を修正
+
+## 0.0.2+2
+
+* Fix: Streams.generate()がlistne前にStreamを終了する場合がある問題を修正
+
+## 0.0.2+1
+
+* Add: Streams.generateのコールバックをStreamEmitインターフェースに変更
+
+## 0.0.2
+
+* Chg: flutter 3.0
+* Add: Streams.generate()を追加
+* Del: Tupleライブラリを削除して標準機能に変更
+
 ## 0.0.1
 
 * Beta Release.